:wrench: Moved category steps into right directory
