+++ src/components/page/CategorySteps.js
