+++ src/components/page/Pages.js
@@ -4,7 +4,7 @@ import { Link, Redirect, withRouter } from 'react-router-dom'
 import TitleHeader from '../layout/TitleHeader'
 import { getDifficultyColorForTag, urlify } from '../../lib/helpers'
 import CategorySidebar from '../layout/CategorySidebar'
-import CategorySteps from '../layout/CategorySteps'
+import CategorySteps from './CategorySteps'
 import PageContent from './PageContent'
 
 class Pages extends React.Component {
